+++ src/lib/Scenes/Sales/index.tests.tsx
@@ -9,7 +9,7 @@ import { GlobalStoreProvider } from "lib/store/GlobalStore"
 import { Theme } from "palette"
 import { SalesFragmentContainer } from "./index"
 
-jest.mock("../Home/Components/SaleArtworksHomeRail", () => "")
+jest.mock("lib/Components/LotsByArtistsYouFollowRail/LotsByFollowedArtistsRail.tsx", () => "")
 
 it("renders the ZeroState when there are no sales", () => {
   // @ts-expect-error STRICTNESS_MIGRATION --- 🚨 Unsafe legacy code 🚨 Please delete this and fix any type errors if you have time 🙏
