+++ src/lib/Scenes/Sales/index.tsx
@@ -2,6 +2,7 @@ import { OwnerType } from "@artsy/cohesion"
 import { Sales_me } from "__generated__/Sales_me.graphql"
 import { Sales_sales } from "__generated__/Sales_sales.graphql"
 import { SalesQueryRendererQuery } from "__generated__/SalesQueryRendererQuery.graphql"
+import { LotsByFollowedArtistsRailContainer } from "lib/Components/LotsByArtistsYouFollowRail/LotsByFollowedArtistsRail"
 import { PageWithSimpleHeader } from "lib/Components/PageWithSimpleHeader"
 import { Stack } from "lib/Components/Stack"
 import { defaultEnvironment } from "lib/relay/createEnvironment"
@@ -12,7 +13,6 @@ import { screen } from "lib/utils/track/helpers"
 import React from "react"
 import { RefreshControl, ScrollView } from "react-native"
 import { createRefetchContainer, graphql, QueryRenderer, RelayRefetchProp } from "react-relay"
-import { LotsByFollowedArtistsRailContainer } from "../LotsByArtistsYouFollow/LotsByFollowedArtistsRail"
 import { SaleList } from "./Components/SaleList"
 import { ZeroState } from "./Components/ZeroState"
